+++ cpg-concepts/src/main/kotlin/de/fraunhofer/aisec/cpg/graph/concepts/http/HttpClient.kt
@@ -28,7 +28,7 @@ package de.fraunhofer.aisec.cpg.graph.concepts.http
 import de.fraunhofer.aisec.cpg.graph.Node
 import de.fraunhofer.aisec.cpg.graph.concepts.Concept
 import de.fraunhofer.aisec.cpg.graph.concepts.Operation
-import de.fraunhofer.aisec.cpg.graph.concepts.auth.IdentityAccessManagement
+import de.fraunhofer.aisec.cpg.graph.concepts.iam.IdentityAccessManagement
 import java.util.Objects
 
 /** Represents an [HttpClient]. */
